requirements.txt
mendelai-brat-parser
smart-open
